feat: standardize automatic mixed precision in training loss functions to use bfloat16 for improved performance

+++ src/training/contrastive_learning/contrastive_loss/loss.py
@@ -11,7 +11,10 @@ def forward_loss(model, batch, margin=1.0):
     #
     # where y = 1 for similar pairs (pull together) and y = 0 for
     # dissimilar pairs (push apart beyond the margin).
-    with torch.cuda.amp.autocast(enabled=torch.cuda.is_available()):
+    with torch.cuda.amp.autocast(
+        enabled=torch.cuda.is_available(),
+        dtype=torch.bfloat16,
+    ):
         anchor_emb = encode(model, batch["anchor_input_ids"], batch["anchor_attention_mask"])
         other_emb = encode(model, batch["other_input_ids"], batch["other_attention_mask"])
