+++ src/training/supervised_learning/instruction_fine_tuning/loss.py
@@ -25,7 +25,10 @@ def forward_loss(model, batch):
     #
     # where M_i is the set of assistant-token positions that should contribute
     # to the loss for example i.
-    with torch.cuda.amp.autocast(enabled=torch.cuda.is_available()):
+    with torch.cuda.amp.autocast(
+        enabled=torch.cuda.is_available(),
+        dtype=torch.bfloat16,
+    ):
         logits = model(
             input_ids=batch["input_ids"],
             attention_mask=batch["attention_mask"],
